requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai==1.40.6
python-dotenv==1.0.1
gunicorn==22.0.0
blinker==1.8.2
